added some changes

+++ src/DeLong-Desktop/ApiService/DTOs/Prices/PriceResultDto.cs
@@ -8,7 +8,7 @@ public class PriceResultDto
     public decimal ArrivalPrice { get; set; }  // Kelish narxi
     public decimal SellingPrice { get; set; }  // Sotish narxi
     public string UnitOfMeasure { get; set; } = string.Empty; // kg,dona,karobka,litr
-    public int Quantity { get; set; } // miqdori
+    public decimal Quantity { get; set; } // miqdori
     public ProductResultDto Product { get; set; }
 }
